+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+### Added
+- Implementation of `IndexMut` trait and `get_mut()` function.
+
 ## [1.0.0] - 2025-08-11
 ### Changed
 - Initial release